[xy] Speed up mage init.
